Use VectorXd and MatrixXd instead of VectorXf and MatrixXf in implementation of class NormOne
This modifies NormOne for execution on the CPU

+++ src/hand-tracking/src/cpu/NormOne.cpp
@@ -28,14 +28,14 @@ NormOne::NormOne(const double likelihood_gain) noexcept :
 NormOne::~NormOne() = default;
 
 
-std::pair<bool, VectorXf> NormOne::likelihood(const MeasurementModel& measurement_model, const Ref<const MatrixXf>& pred_states)
+std::pair<bool, VectorXd> NormOne::likelihood(const MeasurementModel& measurement_model, const Ref<const MatrixXd>& pred_states)
 {
     bool valid_agent_measurements;
     Data data_agent_measurements;
     std::tie(valid_agent_measurements, data_agent_measurements) = measurement_model.getAgentMeasurements();
 
     if (!valid_agent_measurements)
-        return std::make_pair(false, VectorXf::Zero(1));
+        return std::make_pair(false, VectorXd::Zero(1));
 
     MatrixXf& measurements = any::any_cast<MatrixXf&>(data_agent_measurements);
 
@@ -57,10 +57,10 @@ std::pair<bool, VectorXf> NormOne::likelihood(const MeasurementModel& measuremen
     }
 
     if (!valid_predicted_measurements)
-        return std::make_pair(false, VectorXf::Zero(1));
+        return std::make_pair(false, VectorXd::Zero(1));
 
 
-    VectorXf likelihood(pred_states.cols());
+    VectorXd likelihood(pred_states.cols());
 
     for (int i = 0; i < pred_states.cols(); ++i)
     {
@@ -68,10 +68,10 @@ std::pair<bool, VectorXf> NormOne::likelihood(const MeasurementModel& measuremen
         for (int j = 0; j < measurements.cols(); ++j)
             sum_diff += std::abs(measurements(0, j) - predicted_measurements(i, j));
 
-        likelihood(i) = static_cast<float>(sum_diff);
+        likelihood(i) = sum_diff;
     }
 
-    likelihood = (-static_cast<float>(pImpl_->likelihood_gain_) * likelihood).array().exp();
+    likelihood = (-(pImpl_->likelihood_gain_) * likelihood).array().exp();
 
 
     return std::make_pair(true, std::move(likelihood));
